feat(jobs): handle multiple build types in Centreon Map jobs.

+++ centreon-build/jobs/map/19.04/des-map-desktop.sh
@@ -36,10 +36,16 @@ if [ -z "$major" -o -z "$minor" ] ; then
 fi
 
 # Recreate p2 directory.
-path="/srv/p2/unstable/$major/$minor/"
-ssh -o StrictHostKeyChecking=no "ubuntu@srvi-repo.int.centreon.com" rm -rf $path
-ssh -o StrictHostKeyChecking=no "ubuntu@srvi-repo.int.centreon.com" mkdir -p $path
-scp -o StrictHostKeyChecking=no -r com.centreon.studio.client.product/target/repository/* "ubuntu@srvi-repo.int.centreon.com":$path
+if [ "$BUILD" '=' 'RELEASE' -o "$BUILD" '=' 'REFERENCE' ] ; then
+  if [ "$BUILD" '=' 'RELEASE' ] ; then
+    path="/srv/p2/testing/$major/$minor/"
+  else
+    path="/srv/p2/unstable/$major/$minor/"
+  fi
+  ssh -o StrictHostKeyChecking=no "ubuntu@srvi-repo.int.centreon.com" rm -rf $path
+  ssh -o StrictHostKeyChecking=no "ubuntu@srvi-repo.int.centreon.com" mkdir -p $path
+  scp -o StrictHostKeyChecking=no -r com.centreon.studio.client.product/target/repository/* "ubuntu@srvi-repo.int.centreon.com":$path
+fi
 
 # Move all installers to an install folder.
 rm -rf ../installs
@@ -49,5 +55,5 @@ cp com.centreon.studio.client.packaging.nsis.x86_64/target/*.exe ../installs/
 cp com.centreon.studio.client.product/target/products/Centreon-Map4.product-macosx.cocoa.x86_64.tar.gz ../installs/
 cd ..
 
-# Copy installers to remote repository
-scp -o StrictHostKeyChecking=no -r installs/* "ubuntu@srvi-repo.int.centreon.com:/srv/p2/unstable/"
+# Copy installers to remote repository.
+put_internal_source "map-desktop" "$PROJECT-$VERSION-$RELEASE" installs/*
